解决合并报告问题
