+++ common/COM_analysis.py
@@ -1,31 +1,9 @@
 from common.COM_data import MyData
-import inspect
 import re
-import sys
 import types
-from case.test_case import *
 from common import COM_utilities
-from common.COM_path import *
 import importlib
-from time import sleep
 from common.COM_path import *
-from airtest.core.api import assert_equal
-from airtest.report.report import simple_report
-from scenes.scenes_login.SCN_gamestart import GameStart
-from scenes.scenes_login.SCN_gameloaded import GameLoaded
-from scenes.scenes_login.SCN_newuser import NewUserGuide
-from scenes.scenes_discover.SCN_discover import Discover
-from scenes.scenes_visualbook.SCN_bookread import BookRead
-from scenes.scenes_visualbook.SCN_bookdetail import BookNewDetail
-from scenes.scenes_login.SCN_signin import SignIn
-from scenes.SCN_pageTurn import PageTurn
-from scenes.scenes_community.SCN_chapteredit import ChapterEdit
-from scenes.scenes_community.SCN_community import Community
-from scenes.scenes_community.SCN_readUGCbook import ReadUGCBook
-
-
-# import allure
-
 
 class MyAnalysis():
     function_regexp = re.compile(r"^\$\{(\w+)\(([\$\w =,]*)\)\}$")
